Jekyll Build Attempted Fix
Something is very wrong with my code! Jekyll seems to be building/serving an old version of the site's css/liquid but with the new text and file names, breaking all links! If I regenerate Jekyll Serve then it corrects itself, so guess I'll try using that as my "Build" for now?